RealTime Log Processor/requirements.txt
kafka-python==2.0.2
streamlit==1.35.0
notify2==0.3.1
watchdog==4.0.0
